Refactor HID out of main USB code
Remove the hacky HID descriptor and HID request handling from the main
USB handler and move it into a separate "driver" interface. This is to
prepare for adding CDC ACM driver to allow us to do logging over USB
instead of needing to go through the debugprobe's UART which just sends
the UART over a CDC ACM interface anyhow. This will be helpful to debug
on actual keyboard and not the breadboard.